# OpenWA/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/channels": {
      "get": {
        "operationId": "ChannelController_findAll",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of subscribed channels"
          },
          "400": {
            "description": "Session not ready"
          }
        },
        "summary": "Get all subscribed channels/newsletters",
        "tags": [
          "channels"
        ]
      },
      "post": {
        "description": "The account becomes the channel owner, which is what makes deleting it possible later — neither engine can delete a channel it does not own.",
        "operationId": "ChannelController_create",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created channel"
          },
          "400": {
            "description": "Session not started, or validation failed"
          },
          "403": {
            "description": "The engine refused — channel creation may be disabled for this account"
          }
        },
        "summary": "Create a channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/sessions/{sessionId}/channels/{channelId}": {
      "get": {
        "operationId": "ChannelController_findOne",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "description": "Channel ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel details"
          },
          "404": {
            "description": "Channel not found"
          }
        },
        "summary": "Get a specific channel by ID",
        "tags": [
          "channels"
        ]
      },
      "delete": {
        "operationId": "ChannelController_unsubscribe",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "description": "Channel ID to unsubscribe from",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully unsubscribed from channel"
          }
        },
        "summary": "Unsubscribe from a channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/sessions/{sessionId}/channels/{channelId}/messages": {
      "get": {
        "operationId": "ChannelController_getMessages",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "description": "Channel ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max messages to return (default 50, max 100)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of channel messages"
          }
        },
        "summary": "Get messages from a channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/sessions/{sessionId}/channels/{channelId}/delete": {
      "post": {
        "description": "Irreversible, and every subscriber loses the channel. Deliberately NOT `DELETE /channels/:channelId` — that route unsubscribes, and the two must not be reachable by the same request with a slip of the wrist. Only the owner can delete.",
        "operationId": "ChannelController_remove",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "description": "Channel ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel deleted"
          },
          "400": {
            "description": "Session not started"
          },
          "403": {
            "description": "The engine refused — not found, or this account does not own it"
          }
        },
        "summary": "Delete a channel this account owns",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/sessions/{sessionId}/channels/{channelId}/mute": {
      "post": {
        "description": "Silences the channel's notifications for this account. The subscription is untouched.",
        "operationId": "ChannelController_mute",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "description": "Channel ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuteChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel muted or unmuted"
          },
          "400": {
            "description": "Session not started, or validation failed"
          },
          "403": {
            "description": "The engine refused"
          }
        },
        "summary": "Mute or unmute a channel",
        "tags": [
          "channels"
        ]
      }
    },
    "/api/sessions/{sessionId}/channels/subscribe": {
      "post": {
        "operationId": "ChannelController_subscribe",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inviteCode": {
                    "type": "string",
                    "description": "Channel invite code (from channel link)",
                    "example": "ABC123xyz"
                  }
                },
                "required": [
                  "inviteCode"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully subscribed to channel"
          }
        },
        "summary": "Subscribe to a channel using invite code",
        "tags": [
          "channels"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.13.0",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "OpenWA instance",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "CreateChannelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Channel name",
            "example": "Product updates",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "Channel description",
            "example": "Release notes and downtime notices"
          }
        },
        "required": [
          "name"
        ]
      },
      "MuteChannelDto": {
        "type": "object",
        "properties": {
          "mute": {
            "type": "boolean",
            "description": "True to mute the channel, false to unmute it. Subscription is unaffected either way.",
            "example": true
          }
        },
        "required": [
          "mute"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}